requirements.txt
sentence-transformers 
scikit-learn
pandas 
numpy 
streamlit
faiss-cpu